Add helper functions for opening and closing the buffer
We have basically the same code for opening and closing the buffer in the
oneshot and in the continuous sample mode path. Factor this out into common
helper functions. Also open the buffer before trying to configure it. This acts
as a kind of mutex to protect one application from messing with another
applications config while it is actively sampling.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>