Add an option to force load a plugin
By setting the OSC_FORCE_LOAD environment variable it is possible to load a
plugin even though it's identify function returns false. This is useful for
testing. OSC_FORCE_LOAD can either contain a space separated list of plugins to
load or the string "all" in which case all plugins will be loaded.

E.g.
	OSC_FORCE_PLUGIN=fmcomms1 ./osc

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>