Re-add support for displaying the sampling frequency
Ever since we added device and channel selection support we didn't had support
for reading the sampling frequency. This patch re-adds support for it. The
sampling frequency will either be read from the device itself or if it doesn't
report a sampling frequency the frequency will be read from the assigned
trigger.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>