Let the plugins save/restore their data to/from the INI file
This is an experimental change aimed at providing a faster loading of
the plugins and the multiosc, when using the network backend of libiio.

Each plugin receives a filename as argument to their init() function,
where they can restore their state from. They also receive a filename as
argument to their destroy() function, so that they can save their state
if they need to.

Signed-off-by: Paul Cercueil <paul.cercueil@analog.com>