datafile_in: Make sure that the buffer size is 8 bytes aligned
The databus is 8 bytes wide which means the transfer buffer needs to be 8 byte
aligned. In 1 TX mode it can happen that the buffer size is only 4 bytes
aligned, if this happens double the buffer size and send the same samples twice
this will make sure that the buffer size is 8 bytes aligend.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>