fmcomms2: Properly count the number of buffer channels
Don't try to guess the number of buffer channels based on the relationship
between the total number of channels and some magic number. Rather properly
count the number of buffer channels.

Since this method unfortunately is broken for older kernel version go back to
hardcoding the number of channels there.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>