Split the OSC between a library, libosc, and the osc executable
When using Glibc or other C libraries available on Linux, GCC will
gladly link a shared library even if some symbols are unresolved (and
will be resolved at runtime).

But under Windows it does not work like that; the DLLs must have all
their symbols resolved. The simplest solution here, is to move all the
public symbols of the OSC code to a library, libosc, that the plugins
can link against.

Signed-off-by: Paul Cercueil <paul.cercueil@analog.com>