OSC: When writing to file, only save one buffer of samples
Even if we have more samples than what we display on the plot, it makes
sense to only save the buffer of samples that is plotted.

Besides, when the trigger is enabled the samples are likely to be moved
around in memory, and we cannot guarantee that the samples located after
the end of the first buffer will be consecutive. This previously
resulted in the generated CSV files sometimes containing discontinuities
in the contained values.

Signed-off-by: Paul Cercueil <paul.cercueil@analog.com>