Fix, clean and enhance the ini_unroll function
It was previously adding random characters at the end of lines located
inside a <SEQ> section.

The handling of <SEQ> changed a bit: the parameters are now supposed to
be "signed long long int" and not doubles. This is simpler to handle and
those parameters were never used as doubles anyway.

Signed-off-by: Paul Cercueil <paul.cercueil@analog.com>