OSC: Dynamically increase the buffer size
iio_buffer_refill does not ensure that the buffer will contain the
number of samples requested. When using the network backend, it is
possible that the buffer will contain samples for channels that we're
not interested into.

What we do here, is double the buffer size in case we detect that there
are less usable samples in the buffer than what we asked for.

Signed-off-by: Paul Cercueil <paul.cercueil@analog.com>