oscplot.c: handle the "cycle" token from INI as nb of microseconds
Previously, it was representing the number of GTK events from the event
list to handle before resuming the standard operation. If no events
were in the list, then no time was spent at all. This caused most of the
test profiles to fail when executed on faster machines.

Now, the "cycle" number represents a number of microseconds that we will
wait before resuming the standard operation.

Signed-off-by: Paul Cercueil <paul.cercueil@analog.com>