plugins/fmcomms2.c: Avoid segfault when unloading the plugin
This fixes a nasty bug that was triggered when the plugin was unloaded.
As the handle to the shared library was closed, the timeout set up to
execute a function every 500ms would trigger a segmentation fault while
trying to execute the callback function, as it wouldn't exist anymore.

This is part of the fix for issue #27.

Signed-off-by: Paul Cercueil <paul.cercueil@analog.com>