plugins/fmcomms1.c: Fix gain lock not always working
steps to reproduce:
Set Lock Channel to on and then change channel 1 and 2 gains to different values (e.g. 20). After this set Lock Channel to off and then change only channel 1 gain to 4.5. Now change channel 2 gain to 4.5. The value of channel 2 does not get written to the iio attribute.

Signed-off-by: Dan Nechita <Dan.Nechita@analog.com>