dmm: Be consistent about what is a valid channel for the DMM
There are several places throughout the DMM where we decide whether a
channel should be displayed in the DMM or not. The criteria used to decide
this slightly differ from place to place.

Add a helper function which is used whether a channel should be displayed
in the DMM and use it throughout the DMM plugin. This gives us consistent
behaviour.

Without this we might get to see a device in the device list, but end up
with having no selectable channels for the device.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>