osc: Handle trigger "sampling_frequency" attribute
Currently the IIO Scope assumes that the attribute that stores the sampling
frequency is the "frequency" attribute for triggers.

This is only true for some old non-ABI compliant drivers. The ABI compliant
drivers use the "sampling_frequency" attribute, so make sure we support it.

Support for "frequency" will be dropped eventually after a reasonably long
transition period has passed.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>