oscplot: Fix detection if device needs trigger
libiio has a long standing bug that causes iio_device_get_trigger() when
the device can have a trigger but no trigger is currently assigned.

The IIO Scope used to work around this, but the workaround was broken in
commit a747d45b1785 ("OSC: Rework handling of right-click in the
device/channel tree"). Bring the workaround back until libiio is fixed and
a long enough transition period has passed.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>