osc: Properly handle the case where no trigger is assigned to a device
If a device supports triggers, but no trigger is assigned
iio_device_get_trigger() is supposed to return 0 and set trigger to NULL.

Unfortunately due to a bug in libiio this got broken at some point and the
IIO Scope started to workaround the broken libiio implementation, but does
no longer handle the correct implementation. Make sure that for now the IIO
Scope can handle both the broken and non-broken libiio versions correctly.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>