trigger_dialog: Update device sampling frequency when no trigger is assigned
Update the device frequency in both cases, when a new trigger gets assigned
to the device as well as when no trigger gets assigned to the device.
Otherwise we'll keep working with the old frequency when the device had a
trigger assigned before, but none is assigned after the update.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>