osc.c: save remote IP address to ini file
This allows for reconnecting to the same machine (given it has the same
IP address) when the profile is reloaded (osc is closed and relaunched).

If the network address is unreachable, the connection process should
generally timeout within a few seconds allowing for the regular manually
network connection method to be used.