Backtrace: Capture the backtrace at the crash moment
When a crash occurs a handler will obtain the stack trace and send it to another process for processing. The handler contains only Async-signal-safe functions. The other process is a child of the main process where non Async-signal-safe functions can be used to process the crash data.
The resulting crash report is printed to file.

The -rdynamic linker flag should help obtaining a more detailed backtrace.

Signed-off-by: Dan Nechita <Dan.Nechita@analog.com>