build: iterate pkg-config dependencies
Running `pkg-config dep1 dep2 dep3` is all-or-nothing operation.  If any of
the deps is not installed (example `dep3`), the CFLAGS and LDFLAGS for the
`dep1` and `dep2` will not be available either.

This could be confusing, as a user might end up looking at dep1 when really
he/she should be looking at dep3

For the case here, if any of the deps is missing the first error is
```
oscmain.c:2:10: fatal error: glib.h: No such file or directory
 #include <glib.h>
```
because the CFLAGS for glib2 are not generated.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>