build: check dependencies before doing anything further
Add a build rule that iterate over all dependencies to make sure that
they're available and notify the user which are missing.  The error is
highlighted in red, so it should be fairly visible to the user.

The `check_deps` build rule has been added as a dependency to other build
rules that deal with compilation & linking to make sure.  This should allow
for some parallelization of build if the user wants to run the build with
multipe jobs, and also make sure that dependencies are checked before doing
anything further.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>