oscplot: Always use line plots for frequency domain data
The frequency domain plot does not have an option to select between lines
and points. Currently when creating a frequency domain plot it will use
the points/lines setting that was selected for the time domain plot.

Make sure that the time domain points/lines setting is ignore for frequency
domain plots and make them always use lines.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>