build: rework dependency checks
Fixes: https://github.com/analogdevicesinc/iio-oscilloscope/issues/74

The `check_deps` dependency would trigger re-compilation of files, even
when not needed.
This fixes this, by converting the `check_deps` into a function/routine
that gets called before the compilation of each .c file.

That way make can rely only on file timestamps when re-building a file, or
more.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>