plugins/scpi: remove argument in scpi_rx_setup() call
Fixes https://github.com/analogdevicesinc/iio-oscilloscope/issues/71

The `scpi_rx_setup()` function does not take any args, but it's called
with one.
The compiler does not seem to complain, especially when checks are a bit lax.
And even though this is not a problem (functionally), it is incorrect (a
bug). This fixes it.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>