daq2: Fix DAC data alignment
The daq2 platform uses a 128-bit transmit data bus and hence requires a 16 byte
size alignment on transmit waveform data.

The default alignment is 8 bytes. If a waveform is only 8 bytes aligned and not
16 byte this will cause the last 8 bytes of the waveform to be truncated on the
daq2 platform causing a discontinuity.

To fix this properly configure the required alignment in the daq2 plugin.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>