oscplot: Rework marker label generation
Currently when creating a marker with an index the label is created by
calling sprintf() on a temporary buffer and then copying the temporary
buffer to the label buffer.

Re-work things so that label is directly sprintf()'ed into the label buffer
without having to use the extra temporary buffer.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>