osc: Fix capture_process() signature
capture_process() is used as a callback function for g_timeout_add_full().
g_timeout_add_full() expects that the function has a void * parameter. Update the
signature of capture_process() to meet this requirement and then also
remove the cast that is no longer necessary.

This fixes a warning that can be seen with gcc-8.

	osc.c:1325:70: error: cast between incompatible function types from
				   ‘gboolean ()(void)’ {aka ‘int ()(void)’} to ‘gboolean
				   (*)(void )’ {aka ‘int ()(void *)’}
				   [-Werror=cast-function-type]

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>