dac_data_manager: Ignore locale when parsing data files
The dac_data_manager module currently uses sscanf() to parse data files.
This has the problem that sscanf() takes the current locale into account
when parsing numbers like doubles. Depending on the local the decimal point
can either be a dot(.) or a comma(,).

This means that for locales that use a comma for the decimal point data
files which use a comma for separating the values on a line the file does
get parsed incorrectly and can not be loaded.

Fix this by using g_ascii_strtod() function to parse the values. It parses
the value as if the locale was "C". This means dots are used for the
decimal point.

Note that while there are locale independent variants of sscanf() each
platform has their own incompatible version. Hence we are creating our own
tokenizer that splits the line into its values.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>