treewide: remove re-assignments reported by codacy
Some values are assigned init values, then re-assigned actual values.
Codacy reports them as coding style issues, and while the compiler would
optimize them out, it's a good idea to also do these in the code.

Some re-assignments are computed values, and it's good to not rely on the
compiler to optimize them out.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>