plugins/scpi.c: simplify timeout computation in tty_read()
This was caught due to the double assignment of `ts_end.tv_nsec`.

Technically, we don't need to do that much computation for `ts_end.tv_sec`
because it just wastes computation cycles by doing a 10^9, then dividing by
10^9, just to add 2 seconds to `ts_current.tv_sec`.

The nanoseconds would eventually just get discarded by the division, so
there is no need to consider them in the computation of seconds.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>