Replace snprintf() + fwrite() with fprintf()
Replace the sequence of

	snprintf(buf, ...);
	fwrite(buf, ..., f);
with
	fprintf(f, ...);

This is semantically equivalent for the most part but removes the need for
a temporary buffer. This new solution differs in that the data written is
not truncated if it exceeds the size of the original temporary buffer.

Bulk of the transformation was done using the following coccinelle semantic
patch:

// <smpl>
@@
expression buf;
expression len;
expression f;
@@
+fprinf(f,
-snprintf(buf, len,
 ...);
-fwrite(buf, ..., f);
// </smpl>

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>