dialogs.c: handle empty GIT_COMMIT_TIMESTAMP & GIT_VERSION
This can happen when the source code is downloaded as a zip archive from
Github.
For empty GIT_COMMIT_TIMESTAMP, 0 is returned by
`date_compare_against_build_date()`. We can't ascertain if the version on
Github is newer than the one currently built, because there is no git data.

And for the GIT_VERSION definition, an empty string is used.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>