dac_data_manager: Consider device buffer length alignment requirements
DMA buffers typically have an alignment requirement for the buffer length.
Currently this alignment requirement is hardcoded for each platform in the
corresponding plugin.

Newer versions of the kernel will report the alignment requirements
associated with a device through the `length_align_bytes` property. If this
property is present read it and use its value instead of the hardcoded
value from the plugin.

This will allow the DAC data manager to adopt to changes in the alignment
requirements without having to update the plugin.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>