.travis.yml: drop support for CentOS 6
This is in it's own commit to highlight this issue.
The problem is that the glib2-devel library is too old and the build fails
with:
```
[  2%] Building C object CMakeFiles/osc.dir/osc.c.o
cc1: warnings being treated as errors
/docker_build_dir/osc.c: In function 'load_plugins':
/docker_build_dir/osc.c:953: error: implicit declaration of function 'g_thread_new'
/docker_build_dir/osc.c:953: error: assignment makes pointer from integer without a cast
make[2]: *** [CMakeFiles/osc.dir/osc.c.o] Error 1
make[1]: *** [CMakeFiles/osc.dir/all] Error 2
make: *** [all] Error 2
```

I guess we have to draw the line somewhere with what we support, build and
CentOS 6 is a bit too much work to support.
We already have to build `gtkdatabox` for both CentOS 6 & 7, and if we also
need to build glib2 [which is pretty], then it's not worth supporting it
(yet).

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>