.travis.yml: drop build for Xcode 8
Seems to take up to 20-40 minutes to build.
The build also mentions:
```
Warning: You are using macOS 10.11.
We (and Apple) do not provide support for this old version.
```

and
```
Warning: Your Xcode (8.0) is outdated.
Please update to Xcode 8.2.1 (or delete it).
Xcode can be updated from the App Store.
```

No reason to bother with this much.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>