osc: assume a file regular if d_type is 0
On some file-systems, the `d_type` may be invalid/un-populated.
These mixes are a bit rare/esoteric, and [in such cases] we can assume
that the files are regular, because the information is invalid.

Normally these bugs would be in the filesystem or libc, but it's trivial to
handle them here as well.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>