plugins/adrv9009: Refactor plugin to work with multiple adrv9009
The intention is that the plugin UI will scale when it will detect multiple devices (adrv9009-phy, adrv9009-phy-b, etc.).
Most widgets correspond only to a single device but some widgets are shared between all devices.
When only a single device, the plugin should look and work exactly the same as the version before this refactor.

Signed-off-by: Dan Nechita <dan.nechita@analog.com>