oscplot: fix memory corruption in plot_setup()
When initializing a FFT plot, plot_setup initializes the transform
as well as the X and Y axis and passes the handles to the gtkdatabox library.
After initialization osc_plot_update_rx_lbl is called which calls transform_setup,
invalidating the handles sent to the gtkdatabox. This causes gtkdatabox to apply
operations on invalid memory areas. To fix this, we make sure we don't reinitialize
the transform on initial setup.

Signed-off-by: Adrian Suciu <adrian.suciu@analog.com>