osc: Fix trigger corrupting all data when capturing an odd nb of samples
When substracting 'info->offset' from 'offset' when need to have apples and apples. 'offset' is bytes while 'info->offset' is float. So first we divide 'info->offset' to 4 in order to get a quatter of the value (be aware that this is integer division) and then we convert it to bytes.
The issue is that when given an odd number for the sample count, the 'info->offset' gets set to 2 * sample_count which is the value that is being assigned to offset. The value is even but not a multiple of 4 (as it would be if the sample count wast to be an even number) and that is a problem since offset will be used to memcopy from an arrray of floats. Since it's not a multiple of 4, the memcopy will start copying half of floats (sort of speak), half a float from one float value and half from the next float value.

Signed-off-by: Dan Nechita <dan.nechita@analog.com>