oscplot: Don't show the markers that were valid on the previous frame only
Meaning that on the current frame are no longer valid.
In case the captured signal changes so that there are fewer peaks detected than the number of peaks requested by user, it's better to set the peaks to some default value (x=0, y=0) than to show the peaks from the previous capture.

Signed-off-by: Dan Nechita <dan.nechita@analog.com>