build: config.h: Don't hardcode PREFIX path
It is plausible to run cmake with -DCMAKE_INSTALL_PREFIX so that, the
default prefix path (/usr/local) is not used. This is usefull in some
distros like ARCH linux where, /usr/local is not a default path for
looking for shared libraries. Hence, after default build and
installation, we cannot run osc:

"osc: error while loading shared libraries: libosc.so.0: cannot open
shared object file: No such file or directory".

This change makes it possible to install osc in another locations
(eg: /usr/bin) and still be capable of loading/finding all the plugins.

Signed-off-by: Nuno Sá <nuno.sa@analog.com>