build: Fix .desktop entry with pkexec
Some application launchers will execute osc and just reap it so that
osc parent becomes PID 1 which is not allowed by pkexec. Consequently,
osc won't start. The workaround for this is wrapping osc in a script so
that, sh becomes the parent when executing `pkexec osc`. This workaround
can be seen in some other projects (eg: gparted). Some references:

https://bugs.launchpad.net/ubuntu/+source/update-manager/+bug/1020115

Signed-off-by: Nuno Sá <nuno.sa@analog.com>