build: Install adi-osc icons
Some windows-managers render the app icon in the app window. In this
case, a blank icon was coming up for osc. This patch install osc icons
in the hicolor default theme so that all windows come up with a proper
icon (as long as the glade file defines the icon-name).

Signed-off-by: Nuno Sá <nuno.sa@analog.com>