oscplot: Sort device channels naturally before using them
Oscplot needs to display the channels of devices in a natural order.
E.g.
voltage0_i
voltage0_q
voltage1_i
voltage1_q
voltage2_i
...
It has been observed that libiio doesn't alwasy return the channels in a natural order (it not a libiio specification) so we don't rely on it. Instead we sort the channels.

Signed-off-by: Dan Nechita <dan.nechita@analog.com>