glade/oscplot: Covert the file format to Glade v3.8
Glade v3.8 targets GTK2 and so does osc.
Editing glade files with Glade for GTK3 creates inconsistencies. GTK3 has additional properties and considers some GTK2 widgets and properties as deprecated.
A best practice is to use Glade for GTK2 on all glade files and not mix with the other Glade.

Signed-off-by: Dan Nechita <dan.nechita@analog.com>