dialogs.c: update the connect dialog code
We have some new, more feature rich things in the latest version of
libiio (v0.20), so start using them. This makes things a more
discoverable with networking and USB and local all discoverable.

And displays context attributes, which are helpful to make sure you are
talking to the correct board.

Signed-off-by: Robin Getz <robin.getz@analog.com>