osc: Close plugin if init() fails
This patch will close the plugin if it failed to initialize. This means
that the plugin `destroy()` callback will be called and that the plugin is
removed from the plugin list.

The problem is that there is no clear policy on how the plugin should
behave in the `init()` callback if it fails to initialize. Some plugins
might just do some cleaning before returning from init which can be
a problem for plugins as the generic_dac that will try to call
`get_dac_dev_names()` on all the plugins providing one. Some plugins don't
do any cleanup and are just left around for the complete lifetime of the
process or until it is reloaded (since `destroy()` is only called by then).

The natural thing is that if a plugin fails to initialize, it's resources
should just be removed. Also, this patch just sets the policy that a plugin
does not need to, necessarily, do any cleaning in the case of an error
during init.

Being that sad, some plugins might need to be updated for this new behavior
otherwise they can crash during the `destroy()` callback.

Signed-off-by: Nuno Sá <nuno.sa@analog.com>