Implement plugin for CN0540
The CN0540 plugin offers acces to several board functions such as:
	- Voltage buffer control
	- Six channel voltage monitor
	- Sensor voltage offset control
	- Sensor calibration
At startup, the plugin calibrates the sensor. This process consists in
several reads of the input sensor (adc channel) and voltage adjustsments
on the shift voltage side (sensor offset). After the calibration is
done, the measured data should be as close as possible to zero when the
sensor is idle.

Signed-off-by: Cristina Suteu <cristina.suteu@analog.com>
Signed-off-by: Adrian Suciu <adrian.suciu@analog.com>
Signed-off-by: Sergiu Cuciurean <sergiu.cuciurean@analog.com>