osc: Don't call destroy() in failure
If a plugin fails to init, do not call the `destroy()` callback. After
starting to do some plugin cleanups to take into account that a plugin is
now removed from the list of active plugins and that the `destroy()`
callback would be called if the init() fails, I realized that this is not
the best approach. The reason is that it makes it really cumbersome to handle
the `destroy()` callback on the plugin side. We basically need to check
for every element that we need to free, if it was allocated in the first
place because init() can fail at any place and we do not really know
what was already allocated and what wasn't. So, the best approach is to just
call the `destroy()` callback when we want to close the plugin after a
successful init() so that the plugin can make some assumptions. If the init()
callback fails, is the plugin responsibility to clean whatever it has already
allocated...

Signed-off-by: Nuno Sá <nuno.sa@analog.com>