iio_widget: Add new combo box widget
The only difference introduced in the combo box widget is that the
`update_value()` callback is not clearing all the combo box entries at
every call. This assumes that the IIO available attr cannot really change
at runtime which is true most of the times.

Having this done like this, let's us do things like 'widget->save()'
followed by 'widget->update()' in combo boxes signal handlers without
getting an infinite loop (as updating triggers the handler again). With
this, it's not necessary to have workarounds to make sure a save in a
combo box actually worked.

Signed-off-by: Nuno Sá <nuno.sa@analog.com>