iio_widget: Add some helper functions
This patch adds a helper to make widgets updated based on gtk signals.
Almost all plugins are just replicating the same code over and over
again. Hence just make this an iio_widget helper function. Note that
this helper is intended to be used when we want the pass the widget
itself as user data to the gtk signal handler.

The other helper is a typical gtk signal handler that can be passed to
the new `iio_make_widget_update_signal_based()`. Typically, we just want
to call `iio_widget_save()` on the changed widget so that we can add a
wrapper for this.

Signed-off-by: Nuno Sá <nuno.sa@analog.com>