iio_widget: Block gtk signals on widget update
This patch blocks a gtk signal on a widget update. The point is that when
we update a widget, we can receive a different value from the device
from the one in the GUI (eg: a failed call to widget->save() or an
autonomous update). In these case, we don't really want our widget
signal handler to be called on a value that we we know the device is already
holding.

Signed-off-by: Nuno Sá <nuno.sa@analog.com>