osc: Update sampling frequency of ADC when starting a capture
Plugins signal the oscplot as soon as the sampling frequency is changed (usually the control for the sampling frequency is available in the plugin).
However some ADCs might now have an associated plugin but sampling frequency could be changed by using the 'Debug' plugin or even outside of the application. For this scenario, we can't know instantly that the sampling frequency has changed but an improvement is to read the sampling frequency attribute of the ADC right before starting a capture.

Signed-off-by: Dan Nechita <dan.nechita@analog.com>