osc: fix close_plugin() for static plugins
The commit 5ab930e ("osc: Close plugin if init() fails") closes a
plugin if it fails to initialize. However, this was only tested in
dynamically created plugins where there were no issues. However, it
crashes on static plugins because we get our plugin address using
`dlsym()`. Thus, we cannot really access our object after we do a
`dlclose()` on the library.

So, for static plugins, we need to ensure that `dlclose()` is the last
thing we do when closing the plugin. For dynamic plugins, as it was
allocated at runtime, we can (and we must) call `dlclose()` before we
`g_free()` the plugin. Obviously, we cannot free the plugin and then try
to use `plugin->handle` to close the library.

Signed-off-by: Nuno Sa <nuno.sa@analog.com>