plugins: dac_data_manager: fix I and Q frequency mismatch
The goal of this API is to handle the frequency attr change for the I
type channels. In single/dual tone modes, the Q channel is not visible
to the user and the plugin automatically changes the gtk widget causing
its signal to be called and thus, changing the value in the device. The
problem is that for some frequency values, we won't get exactly what we
tried to set (due to integer approximations in the driver) which means
that 'widget->update(widget)' called from 'iio_widget_save()' will trigger
another gtk signal to set the frequency on the Q channel for the value we
got from the driver. Thus, we will end up will slightly different setting
in the I and Q channels. With this handler, we make use of the
'*_block_signals_by_data' to save the widget value and make sure that
'widget->update(widget)' won#t trigger another signal in case we get a
different value.

Signed-off-by: Nuno Sá <nuno.sa@analog.com>