plugins: adrv9002: don't rearm the timer on profile failure
Do not rearm the timer in case the profile fails to load. The plugin was
getting unresponsive in this situation. Moreover, it does not really make
sense to anything on the device if the profiles failed to load. Our only
choice is to try to load the profile again (or some other profile).

Signed-off-by: Nuno Sá <nuno.sa@analog.com>