plugins: adrv9002: support ORx controls
The way ORx is supposed to work is to enable it only if RX is not in
rf_enabled and TX __is__ in rf_enabled state. After that point we should
not really touch any of the RX controls as it might trigger some state
change that could break the Orx capture. For TX, we are also not supposed
to do any state change on the port as some state transitions break ORx and
we can't recover from it without toggling the ORx button. However the
transition calibrated to rf_enabled always re-enables the capture so that,
if we make sure the user cannot touch the ensm and the port_en widgets, we
should be fine in leaving the other controls as the transitions they will
imply are rf_enabled -> calibrated (breaks ORx) to apply the control and
then calibrated -> rf_enabled (re-enables ORx)... And leaving these controls
is helpful to let's say, change a TX track cal and see the effects of it
on ORx port.

Anyways, we need to take care that future firmware releases do
not break this assumption.

Signed-off-by: Nuno Sá <nuno.sa@analog.com>