plugins: adrv9002: check ports state when reloading settings
We should not assume that nothing related to the profile changed when we
hit the reload settings button. This patch also checks the port status
and all the things that might or might not be visible depending on the
profile. While doing it, small improvements were done to the check status
functions.

Signed-off-by: Nuno Sá <nuno.sa@analog.com>