plugins: adrv9002: disable orx_en toggle button
When updating ORx values, we need to disable the orx_en toggle button if we
detect that the Orx port is not enabled. This is needed, for instance, when
loading a profile that does not support Orx. If we had Orx enabled
before updating the profile, we would be left with some TX/RX widgets
blocked... With this, we make sure to call the ORx widget signal and
properly update things.

Signed-off-by: Nuno Sá <nuno.sa@analog.com>