plugins: adrv9002: fix inconsistent labels under Transmit/DDS block
This patch makes sure the dds gets into a consistent block with the rest
of the plugin (dds1 aligned with TX1 and dds2 with TX2). This is done by
checking the next character after the name "axi-adrv9002-tx" (as this is
the pattern we use to look for this devices). If we see a digit, then we
assume that it is the DDS2 and it should be under TX2 settings. If not a
digit, it's TX1. Naturally there's a chance for this to break if the
device name changes in some ways, but we do need to assume some "contracts"
with the name format otherwise this could get more complicated than it
needs to be.

Signed-off-by: Nuno Sá <nuno.sa@analog.com>