plugins: cf_axi_tdd: fix plugin wrong name
After the call to 'get_iio_devices_starting_with()', we get devices sorted
in descending order. So for example, our array will look something like:

 * 0: axi-tdd-2
 * 1: axi-tdd-1

Hence, as we were just going through the array in a ascending order and
setting up the plugin name directly with the loop index, we would end up
with crossed plugin names. This means, that our plugin 'AXI-CORE-TDD-1'
would actually be handling the device 'axi-core-tdd-2' which is obviously
misleading.

This patch fixes this by going through the array in descending order and
making sure we set the correct plugin name based on our iteration index and
the number of available devices.

Signed-off-by: Nuno Sá <nuno.sa@analog.com>