iio_utils: support labels when sorting devices
The way to distinguish between multiple instances of the same device is
to use IIO labels. As drivers will start to enforce that, we need to
make sure that the sorting funtion 'iio_dev_cmp_by_name()' first looks
for the existance of the device label. If we don't do that and just
compare the device name we won't have any sorting as the device name
should be the same for all instances.

Signed-off-by: Nuno Sá <nuno.sa@analog.com>