osc: add helper for dynamic plugins
All plugins supporting dynamic creation are duplicating code.
Hence, this patch introduces an helper to get the possible number of
plugins together with context for each.

On top of providing an unified API for this, this patch also adds
support for IIO labels for identifying multiple instances of the same
device.

Signed-off-by: Nuno Sá <nuno.sa@analog.com>