dialogs: refactor 'connect_fillin()' function
The main refactor here is that if there's no IIO context we just set a
proper message in the text view and return. There's no added value in
going further and adding messages in the other text views. With this,
we do not need to be always checking if 'ctx' is NULL or not.

There's also more stylistic changes in this patch...

Signed-off-by: Nuno Sá <nuno.sa@analog.com>