plugins: lidar: do not use variadic arrays
A 'const inst' is not a compile time expression which means that if such a
variable is used to set the size of an array, the compiler will treat it as
a vla emitting the appropriated code for it. There'n no need for having it
here and vlas are also a bad idea. Moreover, this is a preparation patch
for the next one where the compilation options will be updated and -Wvla
will be added.

Signee-off-by: Nuno Sá <nuno.sa@analog.com>